bucket_to_bucket_template/requirements.txt
functions-framework
google-cloud-storage
google-cloud
google-generativeai
streamlit
